Introduced AppLifecycleHelper
