+++ src/Files.App/Services/UpdateService.cs
@@ -109,7 +109,7 @@ namespace Files.App.Services
 
 		private async Task DownloadAndInstallAsync()
 		{
-			App.SaveSessionTabs();
+			AppLifecycleHelper.SaveSessionTabs();
 			App.AppModel.ForceProcessTermination = true;
 			var downloadOperation = _storeContext?.RequestDownloadAndInstallStorePackageUpdatesAsync(_updatePackages);
 			var result = await downloadOperation.AsTask();
