+++ src/Files.App/ViewModels/Settings/GeneralViewModel.cs
@@ -116,7 +116,7 @@ namespace Files.App.ViewModels.Settings
 		private async void DoRestartAsync()
 		{
 			UserSettingsService.AppSettingsService.RestoreTabsOnStartup = true; // Tells the app to restore tabs when it's next launched
-			App.SaveSessionTabs(); // Saves the open tabs
+			AppLifecycleHelper.SaveSessionTabs(); // Saves the open tabs
 			await Launcher.LaunchUriAsync(new Uri("files-uwp:")); // Launches a new instance of Files
 			Process.GetCurrentProcess().Kill(); // Closes the current instance
 		}
